pyproject.toml
[tool.poetry]
name = "vk-oauth2"
version = "0.1.0"
description = ""
authors = ["Ilia Strelchenya <whoisitlwich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
